refactor: Split software modulation and related tests
